slightly adjust preferred column update timeout
It seems that onDidChangeTextEditorSelection is being fired more often
and/or differently in vscode 1.75.0 so the setTimeout(0) hack is not
long enough to allow the outstanding motions to drain. For now increase
the timeout to try avoid cursor in the wrong position bug while we think
about a better solution.